+++ js/main.js
@@ -1,4 +1,2 @@
-import './mini-pic.js';
-import './big-pic.js';
 import './form.js';
-import './api.js';
+
