+++ js/util.js
@@ -37,7 +37,7 @@ const showAlert = (message) => {
   setTimeout(() => {
     alertContainer.remove();
   }, ALERT_SHOW_TIME);
-}
+};
 
 export {
   getRandomArrayElem,
